refactor(model): move base model utils to new files
