+++ src/scvi/external/poissonvi/_model.py
@@ -24,9 +24,7 @@ from scvi.model._utils import _init_library_size, scatac_raw_counts_properties
 from scvi.model.base import (
     RNASeqMixin,
 )
-from scvi.model.base._utils import (
-    _de_core,
-)
+from scvi.model.base._de_core import _de_core
 from scvi.module import VAE
 from scvi.utils import setup_anndata_dsp
 from scvi.utils._docstrings import de_dsp
