+++ src/scvi/model/base/_rnamixin.py
@@ -17,10 +17,9 @@ from scvi import REGISTRY_KEYS, settings
 from scvi._types import Number
 from scvi.distributions._utils import DistributionConcatenator, subset_distribution
 from scvi.model._utils import _get_batch_code_from_category, scrna_raw_counts_properties
+from scvi.model.base._de_core import _de_core
 from scvi.module.base._decorators import _move_data_to_device
 from scvi.utils import de_dsp, dependencies, unsupported_if_adata_minified
-
-from ._utils import _de_core
 
 try:
     from sparse import GCXS
