+++ tests/model/test_differential.py
@@ -5,12 +5,12 @@ import pytest
 
 from scvi.data import synthetic_iid
 from scvi.model import SCVI
+from scvi.model.base._de_core import _prepare_obs
 from scvi.model.base._differential import (
     DifferentialComputation,
     estimate_delta,
     estimate_pseudocounts_offset,
 )
-from scvi.model.base._utils import _prepare_obs
 
 
 def test_features():
